ARM64: cpu: select HAVE_PBL_MULTI_IMAGES globally
Incoming changes to the macro definition of ENTRY_FUNCTION will result
in removal of barebox_arm_head for ARM64. Besides ENTRY_FUNCTION itself,
only remaining user would be the non-PBL and single-PBL case. We already
mandate PBL for all ARM64 boards, so mandate them to be multi-image
capable as well.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20221024065716.1215046-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>