ARM: i.MX8MM: innocomm-wb15: imply AT803X_PHY for EVK
EVK has Atheros 8035 Ethernet PHY and DHCP times out when using the
generic PHY driver instead of the at803x driver. Link is detected with
both though. Make user lives easier by adding an imply to the PHY driver
when the board is selected. This is chosen instead of a select to allow
users to disable the driver if they use a different base board for the
SoM.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20221024141203.2473619-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>