readline: Complete strings containing whitespaces correctly
Completion strings containing whitespaces have to end up as a single
argv[] argument, thus the whitespaces have to be escaped. Fix this.

Ideally the whitespaces should only be escaped when we are outside of
double or single quotes, but our completion currently doesn't trigger
at all when invokes inside quotes, so we can ignore this case for now.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>