mtd: nand: nand-mxs: Move register definitions to separate file
The GPMI registers are needed in the MTD GPMI driver and also in the
xload-gpmi driver. So far both use their own set of register defines.
Move the MTD GPMI register defines to include/ so that we can use them
in the xload driver as well. While at it replace some register defines
to use FIELD_PREP instead.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>