dma: apbh-dma: Simplify code
The apbh-dma driver supports chaining DMA descriptors while the DMA
engine is running. This is unused in barebox as well as in the Kernel.
We just queue one descriptor chain and wait for its completion, and
the MTD layer doesn't even support the more advanced features. Just
drop the unused code and greatly simplify it.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>