commands: uptime: add note about caveats
Normally timer overrun is not that much of a problem, because there
where it matters, we usually have an is_timeout() in the loop that reads
the timer often enough. This doesn't necessarily hold for the new uptime
command and basically anything can happen between subsequent calls and
if get_time_ns() is not called often enough during that time, we end up
with an inaccurate uptime. Mention this in the extended help text.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20221102085017.1941580-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>