ARM64: asm: mark __barebox_arm64_head as function
Calling objdump -d on the PBL ELF file will not disassemble the
instructions comprising __barebox_arm64_head and instead will just
print a hexdump. This is because we use linker magic to get the section
in front of the text segment and GAS fails to detect that
__barebox_arm64_head too is of function type. This is no longer the case
with ENDPROC as it uses the GAS .type directive to explicitly mark
a symbol as being a function:

  #define ENDPROC(name) \
    .type name, @function; \
    END(name)

This only has effect on the ELF and introduces no functional change for
the objcopied image.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20221125100959.3981619-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>