regulator: rk808: fix misplaced DCDC_REG5
Driver was written to mimic existing stpmic1 regulator driver, which had
the implicit assumption of a 1:1 relationship between of_match_data and
regulator configuration with same indices.

Yet DCDC_REG5 was not at the same place it was in the rk_regulator_cfg
due to likely a copy-paste mistake, leading to possibly the wrong
regulators being controlled. Fix this.

Reported-by: Michael Riesch <michael.riesch@wolfvision.net>
Suggested-by: Michael Riesch <michael.riesch@wolfvision.net>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Reviewed-by: Michael Riesch <michael.riesch@wolfvision.net>
Link: https://lore.barebox.org/20221130171439.2928344-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>