firmware: Makefile: fix firmware file dependency tracking
Commit 663894d99c64 ("firmware: refactor firmware/Makefile") aligned the
barebox code with the kernel code. This commit replaced the usage of
$(fw-external-y) by $(obj-pbl-y) and added the '.gen.o' suffix directly
but didn't adapted the patsubst logic for tracking the binary firmware
files. As result a firmware binary file update is not recognized by make
and and won't make into the final binary. This commit fixes this and so
an update will be recognized automatically without require a 'make
clean' before.

Fixes: 663894d99c64 ("firmware: refactor firmware/Makefile")
Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Link: https://lore.barebox.org/20221201145418.2709053-1-m.felsch@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>