ddr: imx8m: warn about unsupported frequencies
When adding a new board with a yet unsupported DDR frequency, PLL
configuration is silently skipped, leading to a later hang during DDRC
configuration. Improve the bring up user experience by printing a
warning in that case. The user is then expected to extend
imx8mm_fracpll_table, so PBL knows how PLLs should be configured for
this new frequency.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20221205132835.3002605-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>