ARM: dts: fix wrong display-timings native-mode binding
The value of the native-mode property is supposed to be a phandle to
the entry in question, but instead the full node path was used because
the <angle brackets> were missing. Fix this. This should not involve any
functional change as the default mode would've been the first element,
which is the same element the malformed native-mode was meant to point
at anyway.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Reviewed-by: Sam Ravnborg <sam@ravnborg.org>
Link: https://lore.barebox.org/20221221102627.3059068-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>