mci: sdhci: arasan: fix probing eMMC without no-sd
Usual procedure when calling the detect callback of a MCI host driver is
to assume the connected card is a SD-Card and on timeout, treat it as
MMC instead. The Arasan driver failed to do this properly leading to the
need to specify no-sd to get eMMCs working. This is because the driver
only handled software timeouts gracefully: The MCI host itself may also
report timeout and it should not be treated as an error and instead
should be silently passed along as -ETIMEDOUT, so the MCI core can check
against it and fall back to eMMC after a failed probe for SD.

Reported-by: Michael Graichen <michael.graichen@hotmail.com>
Reported-by: Steffen Trumtrar <s.trumtrar@pengutronix.de>
Cc: Yann Sionneau <ysionneau@kalray.eu>
Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Link: https://lore.barebox.org/20230103220317.466767-1-ahmad@a3f.at
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>