ARM: rockchip: make ATF installation optional
On RK3399, ATF running as BL31 will disable normal world access to
some peripherals. To make it easier to develop PBL code that accesses
these peripherals prior to ATF installation, support running barebox
proper in EL3, so all of memory can be poked from the barebox shell.

This is strictly a development feature. Linux on ARM64 is not meant to
be booted without PSCI, which the secure monitor provides.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Link: https://lore.barebox.org/20230109180024.2018960-1-ahmad@a3f.at
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>