ARM: dts: rpi3: Enable sdhci controller
We disabled the sdhci controller in the devicetree because trying
to detect the attached SDIO resulted in annoying error messages.
As we can now gracefully detect SDIO cards we can enable the
controller. This has the advantage that Linux gains WiFi support
when the kernel is started with the barebox live tree.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>
Link: https://lore.barebox.org/20230110155039.2039341-3-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>