ARM: imx6sx: udoo-neo: enable memory size auto detection
Not all udoo-neo board have same memory size. So, remove memory node
to allow memory size auto detection.

Potentially we can use different DTs for different board in the barebox.
But we are able to auto detect board version, better we keep udoo-neo-full
as the basis version and disable not supported parts by the board code
if needed.

Signed-off-by: Oleksij Rempel <o.rempel@pengutronix.de>
Link: https://lore.barebox.org/20230113113407.2684725-6-o.rempel@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>